Fixes compiler warning.
git-svn-id: http://svn.mythtv.org/svn/trunk@7600 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmyth/uitypes.cpp
@@ -2693,7 +2693,7 @@ void UIManagedTreeListType::Draw(QPainter *p, int drawlayer, int context)
             bool selected;
             int pos = current_node->getPosition();
 
-            if (pos > lcddev->getLCDHeight())
+            if (pos > (int)lcddev->getLCDHeight())
                 lnode = nodes->at(pos - lcddev->getLCDHeight());
             else
                 lnode = nodes->first();
